refactor eval types
