+++ packages/database/src/server/models/ragEval/dataset.ts
@@ -1,9 +1,8 @@
+import { RAGEvalDataSetItem } from '@lobechat/types';
 import { and, desc, eq } from 'drizzle-orm';
 
-import { RAGEvalDataSetItem } from '@/types/eval';
-
-import { LobeChatDatabase } from '../../../type';
 import { NewEvalDatasetsItem, evalDatasets } from '../../../schemas';
+import { LobeChatDatabase } from '../../../type';
 
 export class EvalDatasetModel {
   private userId: string;
