+++ packages/types/src/index.ts
@@ -1,6 +1,7 @@
 export * from './artifact';
 export * from './chunk';
 export * from './clientDB';
+export * from './eval';
 export * from './fetch';
 export * from './knowledgeBase';
 export * from './message';
