+++ src/store/knowledgeBase/slices/ragEval/actions/evaluation.ts
@@ -1,10 +1,10 @@
+import { CreateNewEvalEvaluation, RAGEvalDataSetItem } from '@lobechat/types';
 import { SWRResponse, mutate } from 'swr';
 import { StateCreator } from 'zustand/vanilla';
 
 import { useClientDataSWR } from '@/libs/swr';
 import { ragEvalService } from '@/services/ragEval';
 import { KnowledgeBaseStore } from '@/store/knowledgeBase/store';
-import { CreateNewEvalEvaluation, RAGEvalDataSetItem } from '@/types/eval';
 
 const FETCH_EVALUATION_LIST_KEY = 'FETCH_EVALUATION_LIST_KEY';
